Increment de versió, canvi de logo i tractament millorat de error al eliminar un estat.

+++ ripea-service/src/main/java/es/caib/ripea/service/service/ExpedientEstatServiceImpl.java
@@ -1,17 +1,10 @@
 package es.caib.ripea.service.service;
 
-import es.caib.ripea.persistence.repository.ExpedientEstatRepository;
-import es.caib.ripea.persistence.repository.ExpedientRepository;
-import es.caib.ripea.persistence.repository.UsuariRepository;
-import es.caib.ripea.persistence.repository.command.ExpedientRepositoryCommnand;
-import es.caib.ripea.persistence.entity.*;
-import es.caib.ripea.service.helper.*;
-import es.caib.ripea.service.helper.PaginacioHelper.Converter;
-import es.caib.ripea.service.intf.dto.*;
-import es.caib.ripea.service.intf.exception.NotFoundException;
-import es.caib.ripea.service.intf.exception.ValidationException;
-import es.caib.ripea.service.intf.service.ExpedientEstatService;
-import es.caib.ripea.service.intf.utils.Utils;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,15 +14,42 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import es.caib.ripea.persistence.entity.EntitatEntity;
+import es.caib.ripea.persistence.entity.ExpedientEntity;
+import es.caib.ripea.persistence.entity.ExpedientEstatEntity;
+import es.caib.ripea.persistence.entity.MetaExpedientEntity;
+import es.caib.ripea.persistence.entity.UsuariEntity;
+import es.caib.ripea.persistence.repository.ExpedientEstatRepository;
+import es.caib.ripea.persistence.repository.ExpedientRepository;
+import es.caib.ripea.persistence.repository.MetaExpedientTascaRepository;
+import es.caib.ripea.persistence.repository.UsuariRepository;
+import es.caib.ripea.persistence.repository.command.ExpedientRepositoryCommnand;
+import es.caib.ripea.service.helper.ContingutHelper;
+import es.caib.ripea.service.helper.ConversioTipusHelper;
+import es.caib.ripea.service.helper.DateHelper;
+import es.caib.ripea.service.helper.EntityComprovarHelper;
+import es.caib.ripea.service.helper.ExpedientEstatHelper;
+import es.caib.ripea.service.helper.MetaExpedientHelper;
+import es.caib.ripea.service.helper.PaginacioHelper;
+import es.caib.ripea.service.helper.PaginacioHelper.Converter;
+import es.caib.ripea.service.intf.dto.ContingutMassiuFiltreDto;
+import es.caib.ripea.service.intf.dto.ExpedientDto;
+import es.caib.ripea.service.intf.dto.ExpedientEstatDto;
+import es.caib.ripea.service.intf.dto.ExpedientEstatEnumDto;
+import es.caib.ripea.service.intf.dto.PaginaDto;
+import es.caib.ripea.service.intf.dto.PaginacioParamsDto;
+import es.caib.ripea.service.intf.dto.ResultDto;
+import es.caib.ripea.service.intf.dto.ResultEnumDto;
+import es.caib.ripea.service.intf.exception.NotFoundException;
+import es.caib.ripea.service.intf.exception.ValidationException;
+import es.caib.ripea.service.intf.service.ExpedientEstatService;
+import es.caib.ripea.service.intf.utils.Utils;
 
 @Service
 public class ExpedientEstatServiceImpl implements ExpedientEstatService {
 
 	@Autowired private ExpedientEstatRepository expedientEstatRepository;
+	@Autowired private MetaExpedientTascaRepository metaExpedientTascaRepository;
 	@Autowired private ConversioTipusHelper conversioTipusHelper;
 	@Autowired private ContingutHelper contingutHelper;
 	@Autowired private PaginacioHelper paginacioHelper;
@@ -225,8 +245,12 @@ public class ExpedientEstatServiceImpl implements ExpedientEstatService {
 		entityComprovarHelper.comprovarEntitatPerMetaExpedients(entitatId);
 		ExpedientEstatEntity entity = expedientEstatRepository.getOne(expedientEstatId);
 		if (!entity.getExpedients().isEmpty()) {
-			throw new ValidationException("");
+			throw new ValidationException("expedient.estat.controller.esborrat.error.restriccio1");
 		}
+		if (metaExpedientTascaRepository.countByEstatCrearTascaOrEstatFinalitzarTasca(entity, entity)>0) {
+			throw new ValidationException("expedient.estat.controller.esborrat.error.restriccio2");
+		}
+		
 		expedientEstatRepository.delete(entity);
 		
 		if (rolActual.equals("IPA_ORGAN_ADMIN")) {
